missing renaming of main module for physical data model
