+++ cognite/neat/core/_constants.py
@@ -12,7 +12,7 @@ from cognite import neat
 from cognite.neat.core._issues.errors._general import NeatValueError
 
 if TYPE_CHECKING:
-    from cognite.neat.core._data_model.models.dms import PhysicalProperty
+    from cognite.neat.core._data_model.models.physical import PhysicalProperty
 
 
 def _is_in_notebook() -> bool:
@@ -198,7 +198,7 @@ BASE_MODEL = Literal["CogniteCore"]
 
 def get_asset_read_only_properties_with_connection() -> "list[PhysicalProperty]":
     """Gets the asset read-only properties with connection, i.e. Root and Path."""
-    from cognite.neat.core._data_model.models.dms import PhysicalProperty
+    from cognite.neat.core._data_model.models.physical import PhysicalProperty
 
     return [PhysicalProperty.model_validate(item) for item in (_ASSET_ROOT_PROPERTY, _ASSET_PATH_PROPERTY)]
