+++ cognite/neat/core/_data_model/_shared.py
@@ -8,7 +8,7 @@ from cognite.neat.core._data_model.models import (
 from cognite.neat.core._data_model.models.conceptual._unverified import (
     UnverifiedConceptualDataModel,
 )
-from cognite.neat.core._data_model.models.dms._unverified import (
+from cognite.neat.core._data_model.models.physical._unverified import (
     UnverifiedPhysicalDataModel,
 )
 from cognite.neat.core._utils.spreadsheet import SpreadsheetRead
