+++ cognite/neat/core/_data_model/models/physical/_validation.py
