+++ cognite/neat/core/_data_model/models/physical/_verified.py
