+++ tests/tests_unit/test_rules/test_transformers/test_mapping.py
@@ -1,11 +1,11 @@
-from cognite.neat.core._data_model.models.dms import (
+from cognite.neat.core._data_model.models.entities import ContainerEntity, ViewEntity
+from cognite.neat.core._data_model.models.physical import (
     UnverifiedPhysicalContainer,
     UnverifiedPhysicalDataModel,
     UnverifiedPhysicalMetadata,
     UnverifiedPhysicalProperty,
     UnverifiedPhysicalView,
 )
-from cognite.neat.core._data_model.models.entities import ContainerEntity, ViewEntity
 from cognite.neat.core._data_model.transformers import RuleMapper
